Validation and error handling

+++ src/main/java/com/devsuperior/desafio03/services/ClientService.java
@@ -3,10 +3,12 @@ package com.devsuperior.desafio03.services;
 import com.devsuperior.desafio03.dto.ClientDTO;
 import com.devsuperior.desafio03.entities.Client;
 import com.devsuperior.desafio03.repositories.ClientRepository;
+import com.devsuperior.desafio03.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -17,7 +19,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id) {
-        Client client = repository.findById(id).get();
+        Client client = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ClientDTO(client);
     }
 
@@ -43,8 +46,11 @@ public class ClientService {
         return new ClientDTO(entity);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
         repository.deleteById(id);
     }
